Move CalibrateConfig to CalibrationTools
